+++ code/src/ui/graphic/table/registry/ribbon/table-ribbon.tsx
@@ -17,6 +17,7 @@ import Button from "ui/interaction/button";
 import DateInput from "ui/interaction/input/date-input";
 import ColumnToggle from "../../action/column-toggle";
 import { getDisabledDates } from "../registry-table-utils";
+import { buildUrl } from "utils/client-utils";
 
 interface TableRibbonProps {
   path: string;
@@ -46,8 +47,15 @@ export default function TableRibbon(props: Readonly<TableRibbonProps>) {
   const isPermitted = usePermissionGuard();
   const { navigateToDrawer } = useDrawerNavigation();
   const { resetFormSession } = useOperationStatus();
+  const isContractRegistry: boolean = props.lifecycleStage === LifecycleStageMap.PENDING ||
+    props.lifecycleStage === LifecycleStageMap.ACTIVE ||
+    props.lifecycleStage === LifecycleStageMap.ARCHIVE;
+  const isTaskRegistry: boolean = props.lifecycleStage === LifecycleStageMap.OUTSTANDING ||
+    props.lifecycleStage === LifecycleStageMap.SCHEDULED ||
+    props.lifecycleStage === LifecycleStageMap.CLOSED;
   const isBillingStage: boolean = props.lifecycleStage === LifecycleStageMap.ACCOUNT ||
-    props.lifecycleStage === LifecycleStageMap.PRICING;
+    props.lifecycleStage === LifecycleStageMap.PRICING ||
+    props.lifecycleStage === LifecycleStageMap.INVOICE;
 
   const triggerRefresh: React.MouseEventHandler<HTMLButtonElement> = () => {
     props.triggerRefresh();
@@ -121,90 +129,90 @@ export default function TableRibbon(props: Readonly<TableRibbonProps>) {
                       className="w-full sm:w-auto py-3 sm:py-2 text-sm font-medium"
                     />
                   </div>
+                  <div className="sm:w-auto">
+                    <RedirectButton
+                      label={dict.nav.title.invoice}
+                      leftIcon={"request_quote"}
+                      hasMobileIcon={false}
+                      url={Routes.BILLING_INVOICE}
+                      variant={
+                        props.lifecycleStage === LifecycleStageMap.INVOICE ? "active" : "ghost"
+                      }
+                      className="w-full sm:w-auto py-3 sm:py-2 text-sm font-medium"
+                    />
+                  </div>
                 </>
                 )}
               </div>
             </div>
           )}
       </div>
-      <div className="flex justify-between items-end md:gap-2 lg:gap-0 mt-4 flex-wrap">
-        { /* Remove overflow-hidden when in billing stages to allow dropdown to overflow */}
-        <div className={`flex flex-wrap sm:flex-nowrap items-stretch bg-ring rounded-lg border border-border divide-x divide-border ${isBillingStage ? '' : 'overflow-hidden '}`}>
-          {isPermitted("registryFullAccess") &&
-            (props.lifecycleStage === LifecycleStageMap.PENDING ||
-              props.lifecycleStage === LifecycleStageMap.ACTIVE ||
-              props.lifecycleStage === LifecycleStageMap.ARCHIVE) && (
-              <>
-                <RedirectButton
-                  label={dict.nav.title.pending}
-                  leftIcon="free_cancellation"
-                  hasMobileIcon={false}
-                  url={`${Routes.REGISTRY_GENERAL}/${props.entityType}`}
-                  variant={
-                    props.lifecycleStage == LifecycleStageMap.PENDING ? "active" : "ghost"
-                  }
-                  className="text-sm font-medium !rounded-none !border-0"
-                />
-                <RedirectButton
-                  label={dict.nav.title.active}
-                  leftIcon="check_circle_outline"
-                  hasMobileIcon={false}
-                  url={`${Routes.REGISTRY_GENERAL}/active/${props.entityType}`}
-                  variant={
-                    props.lifecycleStage == LifecycleStageMap.ACTIVE ? "active" : "ghost"
-                  }
-                  className="text-sm font-medium !rounded-none !border-0"
-                />
-                <RedirectButton
-                  label={dict.nav.title.archive}
-                  leftIcon="archive_outlined"
-                  hasMobileIcon={false}
-                  url={`${Routes.REGISTRY_GENERAL}/archive/${props.entityType}`}
-                  variant={
-                    props.lifecycleStage == LifecycleStageMap.ARCHIVE ? "active" : "ghost"
-                  }
-                  className="text-sm font-medium !rounded-none !border-0"
-                />
-              </>
-            )}
-          {isPermitted("registryFullAccess") &&
-            (props.lifecycleStage === LifecycleStageMap.OUTSTANDING ||
-              props.lifecycleStage === LifecycleStageMap.SCHEDULED ||
-              props.lifecycleStage === LifecycleStageMap.CLOSED) && (
-              <>
-                <RedirectButton
-                  label={dict.nav.title.outstanding}
-                  leftIcon="pending"
-                  hasMobileIcon={false}
-                  url={`${Routes.REGISTRY_TASK_OUTSTANDING}`}
-                  variant={
-                    props.lifecycleStage == LifecycleStageMap.OUTSTANDING ? "active" : "ghost"
-                  }
-                  className="text-sm font-medium !rounded-none !border-0 "
-                />
-                <RedirectButton
-                  label={dict.nav.title.scheduled}
-                  leftIcon="schedule"
-                  hasMobileIcon={false}
-                  url={`${Routes.REGISTRY_TASK_SCHEDULED}`}
-                  variant={
-                    props.lifecycleStage == LifecycleStageMap.SCHEDULED ? "active" : "ghost"
-                  }
-                  className="text-sm font-medium !rounded-none !border-0 "
-                />
-                <RedirectButton
-                  label={dict.nav.title.closed}
-                  leftIcon="event_busy"
-                  hasMobileIcon={false}
-                  url={`${Routes.REGISTRY_TASK_CLOSED}`}
-                  variant={
-                    props.lifecycleStage == LifecycleStageMap.CLOSED ? "active" : "ghost"
-                  }
-                  className="text-sm font-medium !rounded-none !border-0 "
-                />
-              </>
-            )}
-        </div>
+      <div className={`flex ${isPermitted("registryFullAccess") && (isContractRegistry || isTaskRegistry) ? "justify-between" : "justify-end"} 
+      items-end md:gap-2 lg:gap-0 mt-4 flex-wrap`}>
+        {isPermitted("registryFullAccess") && (isContractRegistry || isTaskRegistry) &&
+          <div className={`flex flex-wrap sm:flex-nowrap items-stretch bg-ring rounded-lg border border-border divide-x divide-border}`}>
+            {isContractRegistry && <RedirectButton
+              label={dict.nav.title.pending}
+              leftIcon="free_cancellation"
+              hasMobileIcon={false}
+              url={`${Routes.REGISTRY_GENERAL}/${props.entityType}`}
+              variant={
+                props.lifecycleStage == LifecycleStageMap.PENDING ? "active" : "ghost"
+              }
+              className="text-sm font-medium !rounded-none !border-0"
+            />}
+            {isContractRegistry && <RedirectButton
+              label={dict.nav.title.active}
+              leftIcon="check_circle_outline"
+              hasMobileIcon={false}
+              url={`${Routes.REGISTRY_GENERAL}/active/${props.entityType}`}
+              variant={
+                props.lifecycleStage == LifecycleStageMap.ACTIVE ? "active" : "ghost"
+              }
+              className="text-sm font-medium !rounded-none !border-0"
+            />}
+            {isContractRegistry && <RedirectButton
+              label={dict.nav.title.archive}
+              leftIcon="archive_outlined"
+              hasMobileIcon={false}
+              url={`${Routes.REGISTRY_GENERAL}/archive/${props.entityType}`}
+              variant={
+                props.lifecycleStage == LifecycleStageMap.ARCHIVE ? "active" : "ghost"
+              }
+              className="text-sm font-medium !rounded-none !border-0"
+            />}
+            {isTaskRegistry && <RedirectButton
+              label={dict.nav.title.outstanding}
+              leftIcon="pending"
+              hasMobileIcon={false}
+              url={`${Routes.REGISTRY_TASK_OUTSTANDING}`}
+              variant={
+                props.lifecycleStage == LifecycleStageMap.OUTSTANDING ? "active" : "ghost"
+              }
+              className="text-sm font-medium !rounded-none !border-0 "
+            />}
+            {isTaskRegistry && <RedirectButton
+              label={dict.nav.title.scheduled}
+              leftIcon="schedule"
+              hasMobileIcon={false}
+              url={`${Routes.REGISTRY_TASK_SCHEDULED}`}
+              variant={
+                props.lifecycleStage == LifecycleStageMap.SCHEDULED ? "active" : "ghost"
+              }
+              className="text-sm font-medium !rounded-none !border-0 "
+            />}
+            {isTaskRegistry && <RedirectButton
+              label={dict.nav.title.closed}
+              leftIcon="event_busy"
+              hasMobileIcon={false}
+              url={`${Routes.REGISTRY_TASK_CLOSED}`}
+              variant={
+                props.lifecycleStage == LifecycleStageMap.CLOSED ? "active" : "ghost"
+              }
+              className="text-sm font-medium !rounded-none !border-0 "
+            />}
+          </div>
+        }
         <div className="flex items-end flex-wrap gap-2 mt-2 md:mt-0">
           {(props.lifecycleStage == LifecycleStageMap.SCHEDULED ||
             props.lifecycleStage == LifecycleStageMap.CLOSED) && (
@@ -218,22 +226,28 @@ export default function TableRibbon(props: Readonly<TableRibbonProps>) {
           {(props.lifecycleStage == LifecycleStageMap.PENDING ||
             props.lifecycleStage == LifecycleStageMap.GENERAL ||
             props.lifecycleStage === LifecycleStageMap.ACCOUNT ||
+            props.lifecycleStage === LifecycleStageMap.INVOICE ||
             props.lifecycleStage === LifecycleStageMap.PRICING) && (
               <Button
                 leftIcon="add"
                 size="icon"
-                tooltipText={dict.action.addItem.replace(
-                  "{replace}",
-                  props.entityType.replace("_", " ")
-                )}
+                tooltipText={props.lifecycleStage === LifecycleStageMap.INVOICE ? dict.action.addInvoice :
+                  dict.action.addItem.replace(
+                    "{replace}",
+                    props.entityType.replace("_", " ")
+                  )}
                 onClick={() => {
                   browserStorageManager.clear();
                   resetFormSession();
-                  navigateToDrawer(Routes.REGISTRY_ADD,
-                    ...(props.lifecycleStage === LifecycleStageMap.ACCOUNT ||
-                      props.lifecycleStage === LifecycleStageMap.PRICING ? [props.lifecycleStage] : []),
-                    props.entityType
-                  )
+                  if (props.lifecycleStage === LifecycleStageMap.INVOICE) {
+                    window.location.href = buildUrl(Routes.REGISTRY_ADD, LifecycleStageMap.INVOICE);
+                  } else {
+                    navigateToDrawer(Routes.REGISTRY_ADD,
+                      ...(props.lifecycleStage === LifecycleStageMap.ACCOUNT ||
+                        props.lifecycleStage === LifecycleStageMap.PRICING ? [props.lifecycleStage] : []),
+                      props.entityType
+                    )
+                  }
                 }}
               />
             )}